backend/requirements.txt
firebase_admin
flask
mock
pyyaml